+++ packages/miniprogram-element/src/component/button.js
@@ -10,7 +10,8 @@ module.exports = {
     }, {
         name: 'type',
         get(domNode) {
-            return domNode.getAttribute('type') || 'default'
+            // 如果使用默认值 default，基础库中会补充 wx-button[type=default]，导致部分样式优先级处理有问题
+            return domNode.getAttribute('type') || undefined
         },
     }, {
         name: 'plain',