+++ packages/mp-webpack-plugin/src/tmpl/page.tmpl.js
@@ -121,6 +121,9 @@ Page({
         // 处理 intersectionObserver 获取
         this.window.$$createIntersectionObserver = options => wx.createIntersectionObserver(this, options)
 
+        // 初始化页面显示状态
+        this.document.$$visibilityState = 'prerender'
+
         init(this.window, this.document)
         this.setData({
             pageId: this.pageId
@@ -135,7 +138,9 @@ Page({
             window: this.window,
             document: this.document,
         }
+        this.document.$$visibilityState = 'visible'
         this.window.$$trigger('wxshow')
+        this.document.$$trigger('visibilitychange')
     },
     onReady() {
         if (this.pageConfig.loadingText) wx.hideLoading()
@@ -143,9 +148,12 @@ Page({
     },
     onHide() {
         global.$$runtime = null
+        this.document.$$visibilityState = 'hidden'
         this.window.$$trigger('wxhide')
+        this.document.$$trigger('visibilitychange')
     },
     onUnload() {
+        this.document.$$visibilityState = 'unloaded'
         this.window.$$trigger('beforeunload')
         this.window.$$trigger('wxunload')
         if (this.app && this.app.$destroy) this.app.$destroy()